+++ src/allegro_font/font.rs
@@ -39,7 +39,7 @@ impl FontDrawing for Core
 		unsafe
 		{
 			let mut info: ALLEGRO_USTR_INFO = mem::uninit();
-			let ustr = al_ref_buffer(&mut info, text.as_ptr() as *i8, text.len() as c_int);
+			let ustr = al_ref_buffer(&mut info, text.as_ptr() as *i8, text.len() as size_t);
 
 			al_draw_justified_ustr(cast::transmute(font.get_font()), *color, x1 as c_float,
 			                       x2 as c_float, y as c_float, diff as c_float, align.get_allegro_flags(), ustr);
@@ -51,7 +51,7 @@ impl FontDrawing for Core
 		unsafe
 		{
 			let mut info: ALLEGRO_USTR_INFO = mem::uninit();
-			let ustr = al_ref_buffer(&mut info, text.as_ptr() as *i8, text.len() as c_int);
+			let ustr = al_ref_buffer(&mut info, text.as_ptr() as *i8, text.len() as size_t);
 
 			al_draw_ustr(cast::transmute(font.get_font()), *color, x as c_float, y as c_float, align.get_allegro_flags(), ustr);
 		}
@@ -87,7 +87,7 @@ impl Font
 		unsafe
 		{
 			let mut info: ALLEGRO_USTR_INFO = mem::uninit();
-			let ustr = al_ref_buffer(&mut info, text.as_ptr() as *i8, text.len() as c_int);
+			let ustr = al_ref_buffer(&mut info, text.as_ptr() as *i8, text.len() as size_t);
 			al_get_ustr_width(self.get_font() as *_, ustr) as i32
 		}
 	}
@@ -122,7 +122,7 @@ impl Font
 		{
 			let (mut x, mut y, mut w, mut h): (c_int, c_int, c_int, c_int) = mem::uninit();
 			let mut info: ALLEGRO_USTR_INFO = mem::uninit();
-			let ustr = al_ref_buffer(&mut info, text.as_ptr() as *i8, text.len() as c_int);
+			let ustr = al_ref_buffer(&mut info, text.as_ptr() as *i8, text.len() as size_t);
 			al_get_ustr_dimensions(self.get_font() as *_, ustr, &mut x, &mut y, &mut w, &mut h);
 			(x as i32, y as i32, w as i32, h as i32)
 		}
